telegram_bot/requirements.txt
astroid==1.5.2
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
cytoolz==0.8.2
Django==1.11.4
django-choices==1.5.0
django-cors-headers==2.1.0
django-extensions==1.8.1
django-filter==1.0.4
django-filters==0.2.1
djangorestframework==3.6.3
ethereum-abi-utils==0.4.3
ethereum-utils==0.4.0
future==0.16.0
honcho==1.0.1
idna==2.6
isort==4.2.5
lazy-object-proxy==1.3.1
lxml==4.1.1
mccabe==0.6.1
numpy==1.12.1
olefile==0.44
pep8==1.7.0
Pillow==4.1.1
psycopg2==2.7.3
py-ecc==1.1.3
pylint==1.7.1
pylru==1.0.9
pysha3==1.0.2
python-telegram-bot==9.0.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
rlp==0.6.0
scikit-learn==0.19.0
scipy==1.0.0
six==1.10.0
toolz==0.8.2
urllib3==1.22
web3==3.15.0
wrapt==1.10.10
